Refactor recent epoch fetch
